Add another example using fs2.Stream
